pyproject.toml
[project]
name = "starter"
version = "0.1"
description = "Backend starter using Python."
authors = [{ name = "Aryan Iyappan", email = "aryaniyappan2006@gmail.com" }]
dependencies = [
    "falcon>=3.1.1",
    "uvicorn>=0.24.0.post1",
    "sqlalchemy>=2.0.23",
    "pydantic[email]>=2.5.2",
    "argon2-cffi>=23.1.0",
    "pydantic-settings>=2.1.0",
    "alembic>=1.12.1",
    "asyncpg>=0.29.0",
    "redis>=5.0.1",
    "celery[librabbitmq]>=5.3.6",
    "jinja2>=3.1.2",
    "user-agents>=2.2.0",
    "humanize>=4.9.0",
    "di>=0.79.2",
    "sanic>=23.6.0",
    "orjson>=3.9.10",
]
requires-python = ">=3.11"
readme = "README.md"
license = { text = "AGPL-3.0" }

[tool.pdm.dev-dependencies]
test = [
    "pytest>=7.4.3",
    "pytest-xdist>=3.5.0",
    "anyio>=4.1.0",
    "pytest-timeout>=2.2.0",
]

[tool.setuptools.packages.find]
where = ["."] # list of folders that contain the packages (["."] by default)
include = [
    "*",
] # package names should match these glob patterns (["*"] by default)
exclude = [] # exclude packages matching these glob patterns (empty by default)
namespaces = false # to disable scanning PEP 420 namespaces (true by default)
